apps/mqtt-broker/bridge/requirements.txt
paho-mqtt>=2.0.0
influxdb-client>=1.38.0
